+++ multisrc/overrides/mangadventure/arcrelight/AndroidManifest.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+          package="eu.kanade.tachiyomi.extension">
+  <application>
+    <activity android:name="eu.kanade.tachiyomi.multisrc.mangadventure.MangAdventureActivity"
+              android:theme="@android:style/Theme.NoDisplay"
+              android:excludeFromRecents="true">
+      <intent-filter>
+        <action android:name="android.intent.action.VIEW"/>
+        <category android:name="android.intent.category.DEFAULT"/>
+        <category android:name="android.intent.category.BROWSABLE"/>
+        <data android:host="arc-relight.com"
+              android:pathPattern="/reader/..*"
+              android:scheme="https"/>
+        <data android:host="www.arc-relight.com"
+              android:pathPattern="/reader/..*"
+              android:scheme="https"/>
+      </intent-filter>
+    </activity>
+  </application>
+</manifest>
