+++ multisrc/overrides/mangadventure/arcrelight/src/ArcRelight.kt
@@ -1,0 +1,27 @@
+package eu.kanade.tachiyomi.extension.en.arcrelight
+
+import eu.kanade.tachiyomi.multisrc.mangadventure.MangAdventure
+
+/** Arc-Relight source. */
+class ArcRelight : MangAdventure(
+    "Arc-Relight",
+    "https://arc-relight.com",
+    listOf(
+        "4-Koma",
+        "Chaos;Head",
+        "Collection",
+        "Comedy",
+        "Drama",
+        "Jubilee",
+        "Mystery",
+        "Psychological",
+        "Robotics;Notes",
+        "Romance",
+        "Sci-Fi",
+        "Seinen",
+        "Shounen",
+        "Steins;Gate",
+        "Supernatural",
+        "Tragedy"
+    )
+)
