+++ multisrc/overrides/mangadventure/default_res/mipmap-hdpi/ic_launcher.png
