+++ multisrc/overrides/mangadventure/default_res/mipmap-mdpi/ic_launcher.png
