+++ multisrc/overrides/mangadventure/default_res/mipmap-xhdpi/ic_launcher.png
