+++ multisrc/overrides/mangadventure/default_res/mipmap-xxhdpi/ic_launcher.png
