+++ multisrc/overrides/mangadventure/default_res/mipmap-xxxhdpi/ic_launcher.png
