+++ multisrc/overrides/mangadventure/default_res/web_hi_res_512.png
