--- src/all/mangadventure/AndroidManifest.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="eu.kanade.tachiyomi.extension">
-
-    <application>
-        <activity
-            android:name=".all.mangadventure.MangAdventureActivity"
-            android:excludeFromRecents="true"
-            android:theme="@android:style/Theme.NoDisplay">
-            <intent-filter>
-                <action android:name="android.intent.action.VIEW" />
-
-                <category android:name="android.intent.category.DEFAULT" />
-                <category android:name="android.intent.category.BROWSABLE" />
-
-                <!-- MangAdventure sites can be added here. -->
-                <data
-                    android:host="arc-relight.com"
-                    android:pathPattern="/reader/..*"
-                    android:scheme="https" />
-                <data
-                    android:host="reader.decadencescans.com"
-                    android:pathPattern="/reader/..*"
-                    android:scheme="https" />
-            </intent-filter>
-        </activity>
-    </application>
-</manifest>
