--- src/all/mangadventure/build.gradle
@@ -1,12 +1,0 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-
-ext {
-    extName = 'MangAdventure'
-    pkgNameSuffix = 'all.mangadventure'
-    extClass = '.MangAdventureFactory'
-    extVersionCode = 8
-    libVersion = '1.2'
-}
-
-apply from: "$rootDir/common.gradle"
